Add factory pattern and spread out filters into own files
